refactor: using folder structure like FSD
