+++ src/pages/auth/Register/RegisterPage.hbs
