+++ src/pages/auth/Register/index.ts
