+++ src/shared/styles/error/error.css
